+++ libs/debugger/include/debugger/RstParser.h
@@ -26,7 +26,7 @@ private:
                                            size_t byteSize);
     std::shared_ptr<EnumType> AddEnumType(const std::string& typeDefId, const std::string& typeName,
                                           bool isSigned, size_t byteSize,
-                                          std::unordered_map<ssize_t, std::string> valueToId);
+                                          std::unordered_map<int64_t, std::string> valueToId);
     std::shared_ptr<ArrayType> AddArrayType(const std::string& typeDefId,
                                             std::shared_ptr<Type> arrayType, size_t numElems);
     std::shared_ptr<StructType> AddStructType(const std::string& typeDefId,
